extracted tracing/profiling/metrics common to observability
