+++ pkg/common/observability/tracing/telemetry.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package common
+package tracing
 
 import (
 	"context"
