+++ pkg/epp/metrics/collectors/inference_pool.go
@@ -20,8 +20,8 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	compbasemetrics "k8s.io/component-base/metrics"
 
+	metricsutil "sigs.k8s.io/gateway-api-inference-extension/pkg/common/observability/metrics"
 	"sigs.k8s.io/gateway-api-inference-extension/pkg/epp/datastore"
-	metricsutil "sigs.k8s.io/gateway-api-inference-extension/pkg/epp/util/metrics"
 )
 
 var (
