+++ pkg/epp/metrics/metrics.go
@@ -27,7 +27,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/metrics"
 
 	logutil "sigs.k8s.io/gateway-api-inference-extension/pkg/common/observability/logging"
-	metricsutil "sigs.k8s.io/gateway-api-inference-extension/pkg/epp/util/metrics"
+	metricsutil "sigs.k8s.io/gateway-api-inference-extension/pkg/common/observability/metrics"
 )
 
 const (
